backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0 